DEV-1923: Extract shared merged-cell undo helpers
Move restoreMergedCells and collectAffectedMergedCells from removeRow.ts
and removeColumn.ts into undoRedo/utils.ts. collectAffectedMergedCells now
takes an axis ('row'|'col') param, removing the duplicated logic SonarCloud
flagged.
